Refactor type hints for AudioBlock, Block, and related modules
